\name{lsp}
\alias{lsp}
\title{List all functions in a package}
\usage{
  lsp(package, all.names = FALSE, pattern)
}
\arguments{
  \item{package}{Which package to list functions from}

  \item{all.names}{Include hidden functions (starting with
  ".")? Defaults to \code{FALSE}}

  \item{pattern}{A regex pattern to limit the results}
}
\description{
  List all functions in a package
}

