% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viridis.R
\name{scale_color_viridis}
\alias{scale_color_viridis}
\alias{scale_fill_viridis}
\title{Viridis color scales}
\usage{
scale_color_viridis(...)

scale_fill_viridis(...)
}
\description{
Uses the viridis color scale.  All other arguments are as to
\link[ggplot2]{scale_color_gradientn}. See \link[viridis]{viridis} for more
information on the color scale.
}

