% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nr.R
\docType{data}
\name{theme_nr}
\alias{theme_nr}
\title{Noam's default ggplot2 styling}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 16.}
\usage{
theme_nr
}
\description{
Noam's default ggplot2 styling
}
\keyword{datasets}
