\docType{data}
\name{theme_nr}
\alias{theme_nr}
\title{Noam's default ggplot2 styling}
\format{List of 13
 $ text              :List of 8
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ panel.grid.major.x: list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.grid.minor.x: list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.grid.minor.y: list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.grid.major.y:List of 4
  ..- attr(*, "class")= chr [1:2] "element_line" "element"
 $ axis.ticks.y      : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.background  : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ legend.title      : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ legend.key        :List of 4
  ..- attr(*, "class")= chr [1:2] "element_rect" "element"
 $ legend.key.size   :Class 'unit'  atomic [1:1] 1.5
  .. ..- attr(*, "unit")= chr "cm"
  .. ..- attr(*, "valid.unit")= int 1
 $ legend.text       :List of 8
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ axis.title        :List of 8
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ axis.text         :List of 8
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 - attr(*, "class")= chr [1:2] "theme" "gg"
 - attr(*, "complete")= logi FALSE}
\usage{
  theme_nr
}
\description{
  Noam's default ggplot2 styling
}
\keyword{datasets}

