% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme_nr.R
\docType{data}
\name{theme_nr}
\alias{theme_nr}
\title{Noam's default ggplot2 styling}
\format{\preformatted{List of 14
 $ text              :List of 8
  ..$ family    : chr "Lato"
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 14
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ panel.grid.major.x: list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.grid.minor.x: list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.grid.minor.y: list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.grid.major.y:List of 4
  ..$ colour  : chr "#ECECEC"
  ..$ size    : num 0.5
  ..$ linetype: num 1
  ..$ lineend : NULL
  ..- attr(*, "class")= chr [1:2] "element_line" "element"
 $ axis.ticks.y      : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.background  : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ legend.title      : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ legend.key        :List of 4
  ..$ fill    : chr "white"
  ..$ colour  : NULL
  ..$ size    : NULL
  ..$ linetype: NULL
  ..- attr(*, "class")= chr [1:2] "element_rect" "element"
 $ legend.key.size   :Class 'unit'  atomic [1:1] 1.5
  .. ..- attr(*, "unit")= chr "cm"
  .. ..- attr(*, "valid.unit")= int 1
 $ legend.text       :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 22
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ axis.text         :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : chr "black"
  ..$ size      : num 13
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ axis.title        :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 22
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ title             :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 22
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 - attr(*, "class")= chr [1:2] "theme" "gg"
 - attr(*, "complete")= logi FALSE
}}
\usage{
theme_nr
}
\description{
Noam's default ggplot2 styling
}
\keyword{datasets}

